# compaction: finish one tree's bar quota before compacting the next