In compaction, at the beginning of every bar, we iterate through each tree and pair of levels (for example Accounts tree and level 3 → 4) and calculate the number of input values to compact. During every _beat_ of the bar, we compact each tree and pair of levels by a fraction of the total input values. This ensures predictability by making sure that compaction does roughly the same amount of work every beat, but also comes with some caveats. Every beat, we:

* Suspend multiple compactions, and carry over their state to the subsequent beat
* Perform multiple instruction-cache context switches to different trees' merge functions, which are type-specialized
* Perform multiple data-cache context switches to different trees' compaction state

This PR changes the way that we schedule compactions. Now, we move on to a subsequent compaction only when the current compaction is run to completion, i.e. when its bar quota is exhausted. This involves pacing compaction using the _total bytes of input values across all trees_, as opposed to using the _number of input values_. This new way of compaction scheduling is not prone to the aforementioned caveats, but also affords us some new optimizations & code simplifications:

* _Memory efficiency_. **Forest compaction now reserves 18MiB of memory for its blocks as opposed to 76.5MiB**! Since we only carry over a single compaction's state to the next beat, we can carry over *more* state! So, we now carry over all the index & data blocks that we read during the current beat, to the next beat. Earlier, we used to discard these blocks and re-read them in the next beat (hopefully from cache, but there was no guarantee).
* We also got rid of some state in compaction, namely _wasted values_ (as a side effect of the above optimization), and also the logic to avoid over-reads at the end of the beat (since over-reads can simply be carried over to the next beat, now). 

---

### Show me the numbers!

#### 1-replica TB cluster (`i4i.16xlarge` EC2 instance)

`./tigerbeetle benchmark --query-count=0 --transfer-count=50_000_000` with our default grid cache size, _~10% increase in throughput, and also a reduction in p50 & p99 latencies_:
```C
load accepted = 182667 tx/s → 201422 tx/s
batch latency p1 = 15 ms → 15 ms
batch latency p50 = 41 ms → 34 ms
batch latency p99 = 184 ms → 177 ms
batch latency p100 = 194 ms → 197 ms
```

`./tigerbeetle benchmark --query-count=0 --transfer-count=50_000_000` with a 64MiB cache to simulate pressure on the grid cache, _~20% increase in throughput, and also a reduction in p50, p99 & p100 latencies_:
```C
load accepted = 164094 tx/s → 196907 tx/s
batch latency p1 = 15 ms → 15 ms
batch latency p50 = 46 ms → 35 ms
batch latency p99 = 184 ms →  175 ms
batch latency p100 = 203 ms → 191 ms
```

#### 6-replica TB cluster (`i4i.xlarge` EC2 instance)

`/tigerbeetle benchmark --query-count=0 --transfer-count=1_000_000_000 --clients=8`, _~3.5% increase in throughput, reduction in p1 & p50 latencies (p100 seems like an outlier here)_

```C
transfer batch size = 8189 txs
transfer batch delay = 0 us
load accepted = 117901 tx/s → 121101 tx/s 
batch latency p1   = 289 ms → 245 ms
batch latency p50  = 557 ms → 554 ms
batch latency p99  = 857 ms → 881 ms
batch latency p100 = 1310 ms → 3122 ms
```