@chaitanyabhandari 2025-06-26
Debugged `./zig/zig build -Drelease vopr -- 6165036051955770325` and it isn't a legitimate liveness issue, the seed passes when ticks-max-convergence is increased from 10_000_000 → 10_010_000. It passes on main with fewer # of ticks, but that wouldn't be a fair comparison since the two runs aren't equivalent (new compaction scheduling that this PR introduces is a source of non-determinism):

```C
./zig/zig build -Drelease vopr -- --ticks-max-convergence=10_010_000 6165036051955770325 (this branch)

 0   \ .        14V 119/123/123C   0:123Jo  0/10J!   0:123Wo <__0:123> v1:1   169Ga  0G!   4G?

          PASSED (10072468 ticks)

./zig/zig build -Drelease vopr -- --ticks-max-convergence=10_010_000 6165036051955770325 (main)

 2   \   .     705V 159/164/164C 133:164Jo  0/_1J! 132:163Wo <__0:__0> v3:3   143Ga  0G!   0G?

          PASSED (2084937 ticks)
```