@matklad 2025-06-26
> 10_000_000

Can we cut that 10x to 1_000_000 ^^ ? While unrelated, it might be useful to learn what takes so much time and optimize that, while we have a convenient repro on hand!