@chaitanyabhandari 2025-06-26
Good idea @matklad, in this particular case I _think_ the root cause could be that the mean write latency is _quite_ high (~1 second), that'd slow things down a fair bit. 
```C
          write_latency_min=3
          write_latency_mean=96
          write_fault_probability=10/100
```