@matklad 2025-07-03
>However, one caveat with this change is that the compaction counters may not be consistent at the end of a beat, they are only consistent at the end of a bar, which is not very interesting since it doesn't track non-determinism. 

On the contrary, it is _very_ interesting! Counters give us a very powerful on-line verification that the entire compaction worked correctly. Internally, compaction is a rather complicated flow. But if you directly count every value that gets in from disk, and every value that gets out to disk, and the numbers of consistent, you have a very strong guarantee that the complicated middle part didn't mess things up! This is beautiful --- we have this very complex ball of asynchrony, and we can check that it worked correctly by just tallying up the water that sucks in and then lets out!

So, I think we should keep the centers, but, yes, it is _crucial_ that we count things as soon as they leave disk (in read_callback), and not at any later time. If we do it later, the counters won't detect a bug where we redundantly read a data block, for example. 

Now, it's a choice whether we want to assert consistency of counters between beats, or only at the end of the bar. I _think_ doing it only at the bar boundary is fine. But we can do this at the beat as well! The "missing" values are exactly those in `level_X_value_blocks`!

So I think we can even strengthen asserts a bit here?

```diff
diff --git a/src/lsm/compaction.zig b/src/lsm/compaction.zig
index afd3627dd..fb5b72297 100644
--- a/src/lsm/compaction.zig
+++ b/src/lsm/compaction.zig
@@ -690,6 +690,9 @@ pub fn CompactionType(
 
                 return 0;
             } else {
+                if (compaction.table_info_a == .immutable) {
+                    compaction.counters.in == compaction.table_info_a.?.immutable.len;
+                }
                 return compaction.quotas.bar;
             }
         }
@@ -813,6 +816,20 @@ pub fn CompactionType(
                 assert(manifest_added_value_count == manifest_removed_value_count);
                 assert(manifest_added_value_count > 0);
             } else {
+                // Cross-check against each over:
+                // - the number of values we should have processed as per original manifest.
+                // - the number of values we processed, accoring to manifest changes.
+                // - individually counted processed values.
+                var values_planned: u64 = switch (compaction.table_info_a.?) {
+                    .immutable => |values| values.len,
+                    .disk => |table_info| table_info.table_info.value_count,
+                };
+                for (compaction.range_b.?.tables.const_slice()) |table_info| {
+                    values_planned += table_info.table_info.value_count;
+                }
+                assert(values_planned == compaction.counters.in);
+                assert(values_planned == manifest_removed_value_count);
+
                 assert(manifest_added_value_count == compaction.counters.out);
                 assert(manifest_removed_value_count == compaction.counters.in);
                 assert(manifest_removed_value_count - manifest_added_value_count ==
@@ -934,8 +951,19 @@ pub fn CompactionType(
             }
 
             assert(compaction.block_queues_empty_output());
-            // We may be carrying over some input blocks to the next beat.
-            maybe(compaction.block_queues_empty_input());
+            var values_in_flight: u64 = 0;
+            for (.{
+                compaction.level_a_value_block.iterator(),
+                compaction.level_b_value_block.iterator(),
+            }) |it| {
+                while (it.next()) |block| {
+                    values_in_flight += Table.data_block_values_used(block.ptr).len;
+                }
+            }
+            values_in_flight -= compaction.level_a_position.value;
+            values_in_flight -= compaction.level_b_position.value;
+            assert(compaction.counters.out + values_in_flight ==
+                compaction.counters.in - compaction.counters.dropped);
 
             assert(compaction.pool.?.idle());
             maybe(compaction.pool.?.blocks_acquired() > 0);
@@ -1446,6 +1474,7 @@ pub fn CompactionType(
             assert(block.stage == .read_value_block);
             stdx.copy_disjoint(.exact, u8, block.ptr, value_block);
             block.stage = .read_value_block_done;
+            compaction.counters.in += Table.data_block_values_used(block.ptr).len;
             compaction.compaction_dispatch();
         }
 
@@ -1555,8 +1584,6 @@ pub fn CompactionType(
 
             compaction.quotas.bar_done += consumed_total;
             compaction.quotas.beat_done += consumed_total;
-            compaction.counters.in += consumed_total;
-
             compaction.counters.dropped += merge_result.dropped;
 
             assert(compaction.quotas.bar_done <= compaction.quotas.bar);
```