@chaitanyabhandari 2025-07-10
New numbers for scale test on a realistic TB cluster! Since the cache here is very large, the optimization in this PR are not _as_ effective. The reason is that our optimizations to carry over all index & value blocks to the next beat only apply if the algorithm of main goes to the _disk_ to re-read those blocks. In case of a large cache, that effect becomes diminished since most blocks that compaction needs (and it has read in the previous beat), are likely to be found in cache! 

Essentially, this PR gives bigger performance gains on _smaller_ TB deployments, which have a smaller cache. For example, I imagine it being helpful for the small serverless tiers!

#### 6-replica TB cluster (`i4i.xlarge` EC2 instance)

`/tigerbeetle benchmark --query-count=0 --transfer-count=1_000_000_000 --clients=8`, with a 24GiB grid cache:  _~3.5% increase in throughput, reduction in p1 & p50 latencies (p100 seems like an outlier here)_

```C
transfer batch size = 8189 txs
transfer batch delay = 0 us
load accepted = 117901 tx/s → 121101 tx/s 
batch latency p1   = 289 ms → 245 ms
batch latency p50  = 557 ms → 554 ms
batch latency p99  = 857 ms → 881 ms
batch latency p100 = 1310 ms → 3122 ms
```